Add an alias for bosh-cli; :open_mouth:
When installing the new Golang Bosh CLI v2 locally, I found that it was
being installed as `gosh` instead of `bosh-cli` or something similar. I
also tend to run `alias gosh=$(which bosh-cli)` in almost every jumpbox
I create. I don't want to completely modify the name of the executable,
but having an alias is great so I can save on typing 27 characters every
time I shell in.

Feel free to close this if it's a big deal, I'll simply go back to typing. :joy: