Attempt to fix issue pasting while SSHed into general-task image
Several people on the team experience issues pasting while SSH'd into a container based on general-task, including our jumpbox containers. Specifically, the pasted sequence is either surrounded by tildes, ~like this~, or the first six characters are either dropped or repeated at random. The bracketed paste control characters are six characters long and if not handled correctly can result in this behavior.

This commit adds an .inputrc to the root user's home directory. Bash will read this file and disable bracketed paste mode. I tested it interactively with `bind 'set enable-bracketed-paste 0'` and it seemed to work.

https://stackoverflow.com/questions/42212099/how-do-i-disable-the-weird-characters-from-bracketed-paste-mode-on-the-mac-os/72071483#72071483